+++ dragonchain/lib/dto/l3_block_model.py
@@ -95,6 +95,8 @@ class L3BlockModel(model.BlockModel):
         self.dc_id = dc_id
         self.current_ddss = current_ddss
         self.block_id = block_id
+        if block_id:
+            self.prev_id = str(int(block_id) - 1)
         self.timestamp = timestamp
         self.prev_proof = prev_proof
         self.scheme = scheme
@@ -116,19 +118,6 @@ class L3BlockModel(model.BlockModel):
     def get_associated_l1_block_id(self) -> Set[str]:
         """Interface function for compatibility"""
         return {self.l1_block_id}
-
-    def export_as_search_index(self) -> Dict[str, Any]:
-        """Export as block search index DTO"""
-        return {
-            "version": "1",
-            "dcrn": schema.DCRN.Block_L3_Search_Index.value,
-            "dc_id": self.dc_id,
-            "block_id": int(self.block_id),
-            "timestamp": int(self.timestamp),
-            "prev_proof": self.prev_proof,
-            "s3_object_folder": "BLOCK",
-            "s3_object_id": self.block_id,
-        }
 
     def export_as_at_rest(self) -> Dict[str, Any]:
         """Export the L3 block that is stored/brodcast/sent for receipt"""